"Removed logger and mapRegistrationRequest from AuthApiController, updated CartController to use AddToCartRequest, added new endpoints to SecurityConfig, updated AddToCartRequest fields to Long and Integer, and commented out logging level in application.properties"

+++ backend/src/main/java/com/springBoot/fruits_ecommerce/models/AddToCartRequest.java
@@ -10,10 +10,11 @@ import lombok.NoArgsConstructor;
 public class AddToCartRequest {
 
     @NotNull(message = "user Id can't be null")
-    private int userId;
+    private Long userId;
 
     @NotNull(message = "product Id can't be null")
-    private int productId;
+    private Long productId;
     @NotNull(message = "quantity can't be null")
-    private int quantity;
+    private Integer quantity;
+
 }
